- StaticCaseCategory use inmemory repository for speed. - Map CaseCategory by id to cirumvent encoding problems.

+++ core-bc/composites/types/src/main/java/se/vgregion/userfeedback/domain/CustomCategory.java
@@ -11,12 +11,12 @@ import java.util.List;
  * @author <a href="mailto:david.rosell@redpill-linpro.com">David Rosell</a>
  */
 @Entity
-public class CustomCategory extends AbstractEntity<Long> implements Serializable {
+public class CustomCategory extends AbstractEntity<Integer> implements Serializable {
     private static final long serialVersionUID = 2103137794928299880L;
 
     @Id
     @GeneratedValue
-    private Long id;
+    private Integer id;
 
     private String name;
 
@@ -28,7 +28,7 @@ public class CustomCategory extends AbstractEntity<Long> implements Serializable
     private List<CustomSubCategory> customSubCategories = new ArrayList<CustomSubCategory>();
 
     @Override
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
